+++ src/main/java/net/silentchaos512/treasurebags/api/IEntityGroup.java
@@ -1,0 +1,12 @@
+package net.silentchaos512.treasurebags.api;
+
+import net.minecraft.entity.Entity;
+import net.minecraft.util.ResourceLocation;
+
+public interface IEntityGroup {
+    ResourceLocation getId();
+
+    boolean matches(Entity entity);
+
+    ResourceLocation getLootTable();
+}